WIP Test: use ~/.cache and add requirements.txt to make Netlify save it
Reading the source code for the build script[1] in Netlify's build
image[2], we can see that the cache directories that are saved and
restored are determined by looking for various files that would
indicate the use of certain tools that would use those caches. Python
apparently uses the ~/.cache/ dir, and this is detected by seeing if
requirements.txt exists.

[1]: https://github.com/netlify/build-image/blob/master/run-build-functions.sh
[2]: https://github.com/netlify/build-image