Test: Add use of _sastack/ pre-built Stack dependencies in some circumstances
If the $NETLIFY_IMAGES_CDN_DOMAIN environment variable is set, try to
use a Stack-supplied toolchain in the _sastack/ (stand-alone Stack)
subdirectory by using an appropriate setting of the $STACK_ROOT
variable. (If _sastack/.local/bin/stack is not executable, this will
immediately fail.)

Use of _sastack/ can be forced with the --netlify-test option, for
local testing and for building that directory.

Currently that directory must be created by hand, but eventually this
should be a pre-existing repo that can be pulled in as a Git submodule
to avoid a long initial build. (Some parts of it may still be rebuilt,
but this shouldn't take long, especially if people commit the rebuilt
version once in a while when it grows large enough.)