Test: Fix trapped exit to preserve error code
Previously, if a command failed with an error code other than 1
causing an exit due to `set -e`, the script would exit with 1 anyway
rather than whatever exit code the failing command returned.

Now we preserve it. This does mean we print "FAILED" yet still exit
with 0 if we don't remember to disable the trap before an early
successful exit. At least there's some indication that there's an
error in the code.

Also, include the error code in the failure message since not
everybody puts the failure code in their next prompt or remembers to
`echo $?` immediately after the command fails.

More than 35 years of programming in Bourne shell and still every
couple of years I find out one of my common idioms is Doing It Wrong.
Why do I still write programs in this language?