WIP Test: Now call _sastack/sastack-setup to configure stack
This adds two new options:
* -b to force building with an sastack-configured Haskell Stack
* -p to test preload of the above; implies -b

When _sastack/sastack-setup is available to execute, or if -b/-p is
given, we rely on sastack-setup to install `stack` if necessary and
configure where it will store its downloads and build products.
Otherwise we just do our own install of `stack` (if necessary) and use
its default configuration.

XXX We're building a cache we want to upload for use by Netlify; it's
not clear where we should be dealing with checking that we're on an
architecture (and maybe distro) that will produce a result compatible
with Netlify.