Test: Ensure full site rebuild after changing the site compiler
If site.hs has changed, triggering a rebuild of the site compiler,
anything and everything in the compiled site could be out of date, so
we need to do a full clean rebuild.

In theory we could check to see if the site compiler was rebuilt, but
this script spends so much time doing other stuff (stack upgrade, site
compiler startup time) that the savings of not doing a site rebuild in
that case are almost nothing, so just take the easy way out and force
a full site rebuild by removing _cache and _site.

We can look again later at speedups for this if developers are
frequently rerunning it, but mostly people changing just the content
(and not the compiler) should be using `site watch` anyway, making
what this script does irrelevant to them.