docroot/: New home for all the static stuff, to keep top dir clean
Everything under docroot/ will get copied to the site as-is (after
removing the docroot/ prefix, of course). Some of this will be stuff
we really do want to be Files We Just Serve (images, presentation
PDFs, maybe even our CSS and JS) and much of it will probably
initially be dumps of stuff from the old site that we eventually want
to generate from some nicer source format.

This does not preclude having other dirs that also have Files We Just
Serve, say, because we want to put all stuff related to a meeting
(announcement, presentations, photographs of it) together in a
different directory.

This also serves as a small example of/exercise in creating a new
route function, `dropInitialComponents`. Useful resources while
writing this were:

* https://jaspervdj.be/hakyll/reference/Hakyll-Core-Routes.html
* http://aherrmann.github.io/programming/2016/01/31/jekyll-style-urls-with-hakyll/
* https://github.com/aherrmann/jekyll_style_urls_with_hakyll_examples/

as well as a bunch of just mechanically taking apart the types and
making sure that what I was writing matched the types, resulting in
the usual Haskell, "if it compiled, it works" situation.

It might though be better to have a `dropInitialDir` function to
provide some error checking ensuring you're dropping a directory, and
not e.g., the filename itself if it's a plain filename.

We could have just hacked this with a `gsubRoute` as done in the
jekyll-style example above:

    dropPostsPrefix :: Routes
    dropPostsPrefix = gsubRoute "posts/" $ const ""

but that doesn't seem right; wouldn't that break on Windows because
it's matching a `/` instead of a `\`?