wiki/: Add wiki pages and code to compile them
We want to keep the source files under `wiki/` using the exact
page names (e.g., `TLUG:MemberGuide`) as used in our dump from
MediaWiki so that any processing code or work at the command line
doesn't need to worry about adding and removing extensions.

This is slightly tricky because we can't directly tell Hakyll the
type of files without extensions. So we copy the code from Hakyll
internals into our own site compiler and do the appropriate
tweaks.

The output files also have no extension; we use a Netlify
configuration (in the newly added `docroot/netlify.toml`) to tell
it to serve these as `text/html`.

This does not yet handle transcludes in pages; the markup is just
dumped straight out into the output file.

The current wiki/ content is just a few sample pages to show that this
is working.