Rename site compiler from `site` to `site-compiler`
This makes it more clear what the program does and also a _lot_
easier to search for places where it's used.

As well as renaming the source file to match, we also move it under
the app/ directory, which is the standard home for Haskell source
files that implement a `main` function for a program.