src/TLUG/{Hello,HelloTest}: HTF test framework and simple demo
We add a TLUG.Hello module with a simple `greet` function and a
corresponding TLUG.HelloTest module with a test for it. This brings
HTF[1], the Haskell Test Framework, which with a few tweaks to
app/Test.hs does the test discovery and instrumentation (with line
numbers) for us. This is based on the HTF tutorial.[2]

When we set `hs-source-dirs: src/` for our test suite in the cabal
file, app/Test.hs was no longer found. Presumably that declaration
replaced a default setting of ./ for the directory in which to look
for sources. We don't actually want to be using ./ as a source
directory to avoid building random other .hs files that might appear
elsewhere in the site, so rather than adding that to `hs-source-dirs`
we instead set `main-is: ../app/Test.hs`, which works fine. I'm not
sure that this is the best way to do this, though.

We also seem to need to add an `other-modules:` declaration to the
cabal test-suite stanza with TLUG.Hello and TLUG.HelloTest. This
doesn't seem scalable, but I'm not sure how we could automatically get
it to use all the necessary modules under src/. Perhaps we need to
build a library against which we'd link both the test suite and the
site-compiler executable (once it starts using library code)?

[1]: https://hackage.haskell.org/package/HTF-0.13.2.5
[2]: https://hackage.haskell.org/package/HTF-0.13.2.5/docs/Test-Framework-Tutorial.html