Test: Make -C clean only explicitly ignored files, not unknown files
All our known build products that we want to clean are explicitly
ignored. Files that are not might be something that will cause
trouble, but they also might be new and wanted source files that
someone just hasn't done a `git add` on yet, in which case removing
them with prejudice would understandably upset developers when they
lose their work.