docroot/netlify.toml: Redirect /wiki and /wiki/ to /wiki/Main_Page
This mostly emulates MediaWiki behaviour. The one difference is that
MediaWiki will issue a 302 temporary redirect from /wiki to /wiki/,
while Netlify seems to detect that /wiki refers to a directory and
just immediately use the /wiki/ redirect. This saves a step and there's
no real need to emulate the MediaWiki two-step behaviour anyway.