TLUG.MediaWiki: Rename variables and reformat
Some variable names have been changed to better match standard Haskell
conventions, and the formatting tweaked to be more readable (to me),
especially when using an 80-column window.

It may be reasonable to completely ignore this commit; minor and
potentially unnecessary refactoring is a technique I often use just to
understand the code.