template/main: Add Netlify site hosting link and badge
This is required so that our `tlug-jp` account on Netlify can be a
"Pro" account, which we need in order to be able to share access to
collaborators rather than sharing a password. It also gives us a
few more features. For details, see:

  https://www.netlify.com/legal/open-source-policy/

There's some issues with our CSS for this; text in the left bar is
made smaller only if it's a link, it seems, so the link text is
smaller than the regular text. (This happens whether it's in the
current <p> block or if it's in a <ul><li> block.)

I chose a <p> block becuase I think this looks better without an
indent; it's not exactly part of the hierarchical lists above, anyway.
But if you're wanting to clean up the formatting of the page, have at
it.