app/Test,.cabal: Restore TLUG.{Hello,HelloTest}
This was commented out (app/Test.hs) and removed (tlug-website.cabal)
a few days ago in commit 8db555c2 ("TLUG.Parser: Refactor parser core
into separate module"), presumably by accident since there was no
mention of it in the commit message.

TLUG.Hello and its test are simple tests originally added to
demonsrate that the test framework actually works. It's an open
question as to whether these are actually still needed, but they do
serve as a handy, simple example and as something to tweak when
playing around with Haskell.