Test: Add install_sysdeps to install pkg-config if not already installed
The Haskell digest-0.0.2.1 apparently requires pkg-config to be
present on the system, and Stack did not detect this and try to
install it for us. (It's not clear whether this is not properly
declared by digest or just that Stack will `apt install` things that
_it_ needs, but doesn't know what dependencies we're trying to build
need.)

We likely didn't encounter this until now because pkg-config is pretty
common, and it's the first time we've ever run it on a system without
that installed.

This assumes a Debian-based system (which is what our Ubuntu automatic
build container is). We may want to consider making this work on
non-Debian-derived distros, but that would be some extra work, and
its likely that anybody running a non-Debian distro and seeing this fail
will understand what to do.