Test install_sysdeps(): Add zlib
We're not required to do this for a Stack build (as far as I've seen),
but if we're trying to a "plain" GHC/Cabal build using a system- or
ghcup-supplied compiler and Cabal, I found that the build failed on
lack of zlib.

This is put here mainly for documentation; it's actually unlikely
that we'd be on any kind of development system that didn't have
the headers etc. for such a commonly used library installed.